xtext: explicit (int) cast on pixel_offset fractional-truncation
Copilot review noted the previous one-line change relied on
implicit double→int truncation when assigning the fractional
pixel offset back to pixel_offset (an int). The behavior was
correct — adj_value - startline is in [0, 1), times fontsize is
in [0, fontsize), truncation toward zero is the desired
rounding — but the intent was buried.

Make it explicit with a leading (int) cast. Same arithmetic,
loud about what's happening, and -Wfloat-conversion doesn't
have to think.

+++ src/xtext.c
@@ -4272,7 +4272,7 @@ gtk_xtext_render_page (GtkXText * xtext)
 	 * fractional remainder — the last line gets clipped even
 	 * after scrolling all the way down. Drop the inner cast
 	 * so the fractional part survives. */
-	xtext->pixel_offset = (gtk_adjustment_get_value(xtext->adj) - startline) * xtext->fontsize;
+	xtext->pixel_offset = (int) ((gtk_adjustment_get_value(xtext->adj) - startline) * xtext->fontsize);
 
 	subline = line = 0;
 	ent = xtext->buffer->text_first;
